Moved trim mode logic to separate hook
